Fix expiry handling. Sending expiry limit unit to seconds

+++ concordium-node/src/plugins/consensus.rs
@@ -65,7 +65,7 @@ pub fn start_consensus_layer(
     ConsensusContainer::new(
         u64::from(conf.maximum_block_size),
         u64::from(conf.block_construction_timeout),
-        u64::from(conf.max_time_to_expiry),
+        conf.max_time_to_expiry,
         u64::from(conf.transaction_insertions_before_purge),
         u64::from(conf.transaction_keep_alive),
         u64::from(conf.transactions_purging_delay),
